% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.main}
\alias{df.main}
\title{Fachkollegien -- Hauptkategorien}
\format{A data frame with 48 rows and 2 variables:
\describe{
     \item{Fach}{Nummer des Fachkollegiums}
     \item{Label}{Name der Nummer des Fachkollegiums}
     }}
\source{
\url{http://www.dfg.de/download/pdf/dfg_im_profil/gremien/fachkollegien/fk-wahl2015/2015_systematik_faecher_fachkollegien.pdf}
}
\usage{
df.main
}
\description{
Fachkollegien -- Hauptkategorien
}
\keyword{datasets}
