% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.sub}
\alias{df.sub}
\title{Fachkollegien -- Unterkategorien}
\format{A data frame with 213 rows and 2 variables:
\describe{
     \item{Fach}{Fachnummer}
     \item{Label}{Name des Fachs}
     }}
\source{
\url{http://www.dfg.de/download/pdf/dfg_im_profil/gremien/fachkollegien/fk-wahl2015/2015_systematik_faecher_fachkollegien.pdf}
}
\usage{
df.sub
}
\description{
Fachkollegien -- Unterkategorien
}
\keyword{datasets}
