% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.dfg}
\alias{df.dfg}
\title{DFG--Kandidierendenliste 2015}
\format{A data frame with 9695 rows and 9 variables:
\describe{
     \item{Fach}{Nummer des Fachkollegiums}
     \item{Person}{Name und Vorname der kandidierenden Person}
     \item{Sex}{Geschlecht der kandidierenden Person}
     \item{Nr}{Fachnummer}
     \item{Institution}{Institution der kandidierenden Person}
     \item{Von}{vorschlagende Institution}
     \item{AK}{Anzahl kandidierender Personen}
     \item{AP}{Anzahl zu waehlender Personen}
     \item{Kategorie}{Kategorisierung der vorschlagenden Institutionen in: Hochschule, Fachgesellschaft, Sonstiges}
}}
\source{
\url{http://www.dfg.de/download/pdf/dfg_im_profil/gremien/fachkollegien/fk-wahl2015/fkwahl_2015_kandidierendenliste.pdf}
}
\usage{
df.dfg
}
\description{
DFG--Kandidierendenliste 2015
}
\keyword{datasets}
