% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.dfg}
\alias{df.dfg}
\title{DFG--Kandidierendenliste 2015 mit Platzierung und Stimmen}
\format{A data frame with 9695 rows and 18 variables:
\describe{
    \item{wb.nummer}{Wissenschaftsbereich Nummer}
    \item{wb.name}{Wissenschaftsbereich Name}
    \item{fg.nummer}{Fachgebiet Nummer}
    \item{fg.name}{Fachgebiet Name}
     \item{fk.nummer}{Fachkollegium Nummer}
     \item{fk.name}{Fachkollegium Name}
     \item{fach.nummer}{Fach Nummer}
     \item{fach.name}{Fach Name}
     \item{person.name}{Name und Vorname der kandidierenden Person}
     \item{person.sex}{Geschlecht der kandidierenden Person}
     \item{person.inst}{Institution der kandidierenden Person}
     \item{inst.von}{vorschlagende Institution}
     \item{inst.von.kat}{Kategorisierung der vorschlagenden Institutionen in: Hochschule, Fachgesellschaft, Sonstiges}
     \item{anzahl.plaetze}{Anzahl zu waehlender Personen}
     \item{anzahl.kandidaten}{Anzahl kandidierender Personen}
     \item{person.platz}{Platzierung der kandidierenden Person}
     \item{person.stimmen}{Anzahl der Stimmen (nur fuer gewaehlte Kandidierende)}
     \item{person.gewaehlt}{Wurde die kandidierende Person gewaehlt? TRUE vs. FALSE}
}}
\source{
\url{http://www.dfg.de/download/pdf/dfg_im_profil/gremien/fachkollegien/fk-wahl2015/fkwahl_2015_kandidierendenliste.pdf}

\url{http://www.dfg.de/download/pdf/dfg_im_profil/gremien/fachkollegien/fk-wahl2015/151126_fkwahl_2015_wahlergebnis_vorlaeufig.pdf}

\url{http://www.dfg.de/download/pdf/dfg_im_profil/gremien/fachkollegien/fk-wahl2015/2015_systematik_faecher_fachkollegien.pdf}
}
\usage{
df.dfg
}
\description{
DFG--Kandidierendenliste 2015 mit Platzierung und Stimmen
}
\keyword{datasets}
